added some basics masks

+++ Agata/Square.h
@@ -19,7 +19,7 @@
 
 */
 
-enum class Square : uint64_t {
+enum class Square : int {
     A1 = 0, B1, C1, D1, E1, F1, G1, H1,
     A2, B2, C2, D2, E2, F2, G2, H2,
     A3, B3, C3, D3, E3, F3, G3, H3,
